Update the regression analysis exceptions (merge with perun 'exceptions', update docstrings etc.)
regression_exceptions.py, exceptions.py: merged the regression exceptions with perun exception module
generic.py, specific.py, methods.py, regression_models.py, tools.py: updated the exceptions module, updated docstrings to reflect changes etc.
